requirements.txt
# FastAPI & server
fastapi>=0.115.0
uvicorn[standard]
python-multipart>=0.0.18
python-dotenv

# Templating
jinja2

# LangChain family (loose but recent stable)
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
langchain-groq>=0.2.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0
# Embeddings
sentence-transformers>=3.0.0
huggingface-hub>=0.25.0

# Qdrant - let pip take latest compatible (1.10+)
qdrant-client>=1.10.0

# PDF - reliable
pymupdf>=1.24.0

# Google direct
google-generativeai>=0.8.0

# ... other packages as before
chromadb>=0.5.0

# Optional
pandas